Change FOV method to use float instead of double

+++ src/main/java/exloran/silvera/mixin/GameRendererMixin.java
@@ -6,15 +6,11 @@ import net.minecraft.client.render.GameRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-/**
- * GameRendererMixin
- * - Zoom için FOV değiştirir
- */
 @Mixin(GameRenderer.class)
 public class GameRendererMixin {
 
     @ModifyReturnValue(method = "getFov", at = @At("RETURN"))
-    private double silvera_modifyFov(double original) {
-        return ZoomModule.getModifiedFov(original);
+    private float silvera_modifyFov(float original) {
+        return (float) ZoomModule.getModifiedFov(original);
     }
 }
